added .isMerged to virtual file

+++ source/lib/fileMixer/fileMixer.write.js
@@ -8,9 +8,22 @@ export default function write(callback) {
 			this.render(done);
 		},
 		(file, done) => {
-			fileSystem.writeFile(file.path, file.contents, error => {
-				done(error, file);
-			});
+			if (file.isFile) {
+				fileSystem.writeFile(file.path, file.contents, error => {
+					done(error, file);
+				});
+			} else {
+				fileSystem.exists(file.path, directoryExists => {
+					if (!directoryExists) {
+						fileSystem.mkdir(file.path, error => {
+							done(error, file);
+						});
+					} else {
+						done(null, file);
+					}
+				});
+			}
+
 		}
 	], callback);
